refactor/ melhora do login e persistencia de pessoa logada
