+++ icasei/src/context/Login.tsx
@@ -1,4 +1,3 @@
-import { useToast } from "@chakra-ui/react";
 import React, { useState } from "react";
 import { useAppNavigate } from "../routes/coordinator";
 
@@ -13,33 +12,23 @@ export const LoginProvider = ({ children }: any) => {
   const [loggedUser, setLoggedUser] = useState<any>();
 
   const { goToSearch, goToLogin } = useAppNavigate();
-  const toast = useToast();
 
   const login = (form: LoginProps) => {
-    let logged = false;
-    localStorage.setItem("logado", form.email);
+    localStorage.setItem("user", form.user);
+    localStorage.setItem("email", form.email);
     setLoggedUser(form);
-    logged = true;
-    if (logged) {
-      toast({
-        title: "Login realizado.",
-        description: `Bem vindo, ${form.user}.`,
-        status: "success",
-        duration: 9000,
-        isClosable: true,
-      });
-      goToSearch();
-    }
+    goToSearch();
   };
 
   const logout = () => {
     setLoggedUser(undefined);
-    localStorage.removeItem("logado");
+    localStorage.removeItem("user");
+    localStorage.removeItem("email");
     goToLogin();
   };
 
   return (
-    <LoginContext.Provider value={{ loggedUser, login, logout }}>
+    <LoginContext.Provider value={{ setLoggedUser, loggedUser, login, logout }}>
       {children}
     </LoginContext.Provider>
   );
