+++ icasei/src/hooks/useUnprotectedPage.ts
@@ -1,12 +1,18 @@
 import React, { useLayoutEffect } from "react";
 import { LoginContext } from "../context/Login";
 
+import { useAppNavigate } from "../routes/coordinator";
+
 export const useUnprotectedPage = () => {
-  const { loggedUser, login } = React.useContext(LoginContext);
+  const { goToSearch } = useAppNavigate();
+  const { login } = React.useContext(LoginContext);
+
+  const localStorageUser = localStorage.getItem("user");
+  const localStorageEmail = localStorage.getItem("email");
 
   useLayoutEffect(() => {
-    if (loggedUser !== undefined) {
-      login(loggedUser);
+    if (localStorageUser?.length && localStorageEmail?.length) {
+      login({ user: localStorageUser, email: localStorageEmail });
     }
-  }, [login, loggedUser]);
+  }, [localStorageUser, localStorageEmail, goToSearch, login]);
 };
